fix rewards page date timezones
the api returns dates in yyyy-mm-dd. we don't actually care
what the exact times are, only that the dates match, so the
rewards page now treats the date string as being in the
browser's time zone. this fixes some issues where, eg, us
time zones would see the date as being off by 1